pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "ura-efris-sdk"
version = "0.1.0"
description = "Python SDK for Uganda Revenue Authority (URA) EFRIS API"
readme = "README.md"
requires-python = ">=3.9"
license = { text = "MIT" }

authors = [
  { name = "Emmanuel Bartile" }
]

keywords = [
  "URA",
  "EFRIS",
  "Uganda",
  "Tax",
  "eInvoicing",
  "Fiscal",
  "SDK"
]

classifiers = [
  "Development Status :: 3 - Alpha",
  "Intended Audience :: Developers",
  "License :: OSI Approved :: MIT License",
  "Programming Language :: Python :: 3",
  "Programming Language :: Python :: 3.9",
  "Programming Language :: Python :: 3.10",
  "Programming Language :: Python :: 3.11",
]

dependencies = [
  "pydantic>=2.0",
  "cryptography>=40.0",
  "pycryptodome>=3.21",
  "pytz>=2023.3",
  "requests>=2.31.0"
]

[project.urls]
Homepage = "https://github.com/paybillke/ura-efris-python-sdk"
Issues = "https://github.com/paybillke/ura-efris-python-sdk/issues"

[tool.setuptools.packages.find]
where = ["."]
include = ["ura_efris_sdk*"]
